+++ packages/theme-patternfly-org/scripts/cli/helpers.js
@@ -1,0 +1,9 @@
+const path = require('path');
+
+function getConfig(options) {
+  return require(path.join(process.cwd(), options.parent.config));
+}
+
+module.exports = {
+  getConfig
+};
