+++ packages/theme-patternfly-org/scripts/cli/start.js
@@ -1,7 +1,11 @@
 const WebpackDevServer = require('webpack-dev-server');
 const webpack = require('webpack');
+const clientConfig = require('../webpack/webpack.client.config');
+const { generate } = require('./generate');
+const { getConfig } = require('./helpers');
+const { watchMD } = require('../md/parseMD');
 
-function start(webpackConfig) {
+function startWebpackDevServer(webpackConfig) {
   const { port } = webpackConfig.devServer;
   const compiler = webpack(webpackConfig);
   const server = new WebpackDevServer(compiler, webpackConfig.devServer);
@@ -11,6 +15,14 @@ function start(webpackConfig) {
       console.log(err);
     }
   });
+}
+
+async function start(options) {
+  generate(options, true);
+  const webpackClientConfig = await clientConfig(null, { mode: 'development', ...getConfig(options) });
+  console.log('start webpack-dev-server');
+  watchMD();
+  startWebpackDevServer(webpackClientConfig);
 }
 
 module.exports = {
